requirements.txt
--extra-index-url https://gitlab.core.rezel.net/api/v4/projects/56/packages/pypi/simple
--extra-index-url https://gitlab.core.rezel.net/api/v4/projects/139/packages/pypi/simple
black<26
fastapi<1
motor<4
netaddr<2
pydantic<3
python-dotenv<2
pyyaml<7
requests<3
uvicorn<1
common-models==0.2.1
rezel-vault-jwt==0.1.1
